fix type

+++ src/algorithms/reco/ScatteredElectronsEMinusPz.cc
@@ -46,7 +46,7 @@ namespace eicrecon {
     // this map will store intermediate results
     // so that we can sort them before filling output
     // collection
-    std::map<double, edm4eic::ReconstructedParticle, std::greater<int>> scatteredElectronsMap;
+    std::map<double, edm4eic::ReconstructedParticle, std::greater<double>> scatteredElectronsMap;
 
     out_electrons->setSubsetCollection();
